add (kludgy) logic to provide for scaling and/or roll-off when D60 sim is used
Support is “hard-coded” and kludgy. It currently only supports D60 simulation within DCI and D65. The scale factor could be automatically determined but would require a self-reference call with a large ACES value to determine the “max CV” that comes out of the math without the scale being applied. This max value is used to determine the proper scale factor such that the maximum color channel scales to 1. At this time, it is unclear if/how roll_white_fwd parameters could be automatically determined.